Refactor into trend chart source text into a component
